Add X-Undelete-Enabled flag for accounts and containers [#85727618]
It can have three states: True, False, or unset (default). Precedence is
from most-specific to least-specific: if the flag is set on the
container, that will be used; otherwise if the flag is set on the
account, that will be used; otherwise, a new global setting
(enable_by_default) will be used.

Only administrators may set the flag, but all users can see the current
state. To unset the flag, POST a value of 'default'.